WDF: improved diode models (#41)
* Update model used for DIodePair

* Adjust diode parameter methods

* Implement diode parameters for single diode

* Clean up diode changes

* {Apply clang-format}

* Fix typos

* Fix MSVC compiler errors

+++ DSP/WDF/r_type.h
@@ -35,8 +35,8 @@ namespace WDFT
         template <typename ElementType, int arraySize, int alignment = 16>
         struct AlignedArray
         {
-            ElementType& operator[] (int index) noexcept { return array[(int) index]; }
-            const ElementType& operator[] (int index) const noexcept { return array[(int) index]; }
+            ElementType& operator[] (int index) noexcept { return array[(size_t) index]; }
+            const ElementType& operator[] (int index) const noexcept { return array[(size_t) index]; }
 
             auto begin() noexcept { return array.begin(); }
             auto begin() const noexcept { return array.begin(); }
